Revert "Add input validation to registration engine (#6659)"
This reverts commit 55fc1c18763cab0602ed9f2d84b8e0b7bbdbcb5f.
