+++ components/registration-framework/org.wso2.carbon.identity.user.registration.engine/src/main/java/org/wso2/carbon/identity/user/registration/engine/graph/UserOnboardingExecutor.java
@@ -83,10 +83,16 @@ public class UserOnboardingExecutor implements Executor {
     public ExecutorResponse execute(RegistrationContext context) throws RegistrationEngineException {
 
         RegisteringUser user = updateUserProfile(context);
-        Map<String, String> userClaims = user.getClaims();
         Map<String, char[]> credentials = user.getUserCredentials();
+
         char[] password =
                 credentials.getOrDefault(PASSWORD_KEY, new DefaultPasswordGenerator().generatePassword());
+
+        Map<String, Object> claims = user.getClaims();
+        Map<String, String> userClaims = new HashMap<>();
+        for (Map.Entry<String, Object> entry : claims.entrySet()) {
+            userClaims.put(entry.getKey(), String.valueOf(entry.getValue()));
+        }
 
         try {
             String userStoreDomainName = resolveUserStoreDomain(user.getUsername());
@@ -164,7 +170,7 @@ public class UserOnboardingExecutor implements Executor {
     private String resolveUsername(RegisteringUser user, String flowId) throws RegistrationEngineException {
 
         String username = Optional.ofNullable(user.getUsername())
-                .orElseGet(() -> Optional.ofNullable(user.getClaims().get(USERNAME_CLAIM_URI)).orElse(""));
+                .orElseGet(() -> Optional.ofNullable(user.getClaims().get(USERNAME_CLAIM_URI)).orElse("").toString());
         if (StringUtils.isBlank(username)) {
             throw handleClientException(ERROR_CODE_USERNAME_NOT_PROVIDED, flowId);
         }
