+++ components/registration-framework/org.wso2.carbon.identity.user.registration.engine/src/main/java/org/wso2/carbon/identity/user/registration/engine/model/RegistrationContext.java
@@ -18,16 +18,14 @@
 
 package org.wso2.carbon.identity.user.registration.engine.model;
 
-import org.slf4j.MDC;
-import org.wso2.carbon.identity.application.authentication.framework.config.model.ExternalIdPConfig;
-import org.wso2.carbon.identity.user.registration.mgt.model.NodeConfig;
-import org.wso2.carbon.identity.user.registration.mgt.model.RegistrationGraphConfig;
-
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
+import org.slf4j.MDC;
+import org.wso2.carbon.identity.application.authentication.framework.config.model.ExternalIdPConfig;
+import org.wso2.carbon.identity.user.registration.mgt.model.NodeConfig;
+import org.wso2.carbon.identity.user.registration.mgt.model.RegistrationGraphConfig;
 
 /**
  * This class is responsible for holding the registration context.
@@ -39,8 +37,6 @@ public class RegistrationContext implements Serializable {
     private final Map<String, String> userInputData = new HashMap<>();
     private final Map<String, Object> properties = new HashMap<>();
     private Map<String, String> authenticatorProperties;
-    private Map<String, Set<String>> currentStepInputs = new HashMap<>();
-    private Map<String, Set<String>> currentRequiredInputs = new HashMap<>();
     private NodeConfig currentNode;
     private RegistrationGraphConfig regGraph;
     private RegisteringUser registeringUser = new RegisteringUser();
@@ -49,7 +45,6 @@ public class RegistrationContext implements Serializable {
     private String userId;
     private String currentActionId;
     private ExternalIdPConfig externalIdPConfig;
-    private Response currentNodeResponse;
 
     public NodeConfig getCurrentNode() {
 
@@ -175,35 +170,5 @@ public class RegistrationContext implements Serializable {
             ExternalIdPConfig externalIdPConfig) {
 
         this.externalIdPConfig = externalIdPConfig;
-    }
-
-    public Map<String, Set<String>> getCurrentStepInputs() {
-
-        return currentStepInputs;
-    }
-
-    public void setCurrentStepInputs(Map<String, Set<String>> currentExpectedInputs) {
-
-        this.currentStepInputs = currentExpectedInputs;
-    }
-
-    public Map<String, Set<String>> getCurrentRequiredInputs() {
-
-        return currentRequiredInputs;
-    }
-
-    public void setCurrentRequiredInputs(Map<String, Set<String>> currentRequiredInputs) {
-
-        this.currentRequiredInputs = currentRequiredInputs;
-    }
-
-    public Response getCurrentNodeResponse() {
-
-        return currentNodeResponse;
-    }
-
-    public void setCurrentNodeResponse(Response response) {
-
-        this.currentNodeResponse = response;
     }
 }
